Refactored filters, added streams
